Revert "SDS Server regression fix (#3192)"
This reverts commit 868e4e590edab82decb1decbd951a56d8817f123.
