+++ changelog/v1.4.0-beta16/upgrade-to-v3-envoy-xds-api-part2.yaml
@@ -1,6 +1,6 @@
 changelog:
 - type: NON_USER_FACING
   issueLink: https://github.com/solo-io/gloo/issues/2815
-  resolvesIssue: false
-  description: >
-    This is the second chunk of work for the upgrade to v3 of the Envoy xDS API. It largely replaces v2 of the listener package with v3 in our code but includes many interdependent changes.
+  resolvesIssue: false 
+  description: > 
+    This is the second chunk of work for the upgrade to v3 of the Envoy xDS API. It largely replaces v2 of the listener package with v3 in our code but includes many interdependent changes.
