+++ projects/gloo/pkg/api/v1/failover.pb.hash.go
@@ -38,6 +38,20 @@ func (m *Failover) Hash(hasher hash.Hash64) (uint64, error) {
 		return 0, err
 	}
 
+	if h, ok := interface{}(m.GetLocality()).(safe_hasher.SafeHasher); ok {
+		if _, err = h.Hash(hasher); err != nil {
+			return 0, err
+		}
+	} else {
+		if val, err := hashstructure.Hash(m.GetLocality(), nil); err != nil {
+			return 0, err
+		} else {
+			if err := binary.Write(hasher, binary.LittleEndian, val); err != nil {
+				return 0, err
+			}
+		}
+	}
+
 	for _, v := range m.GetPrioritizedLocalities() {
 
 		if h, ok := interface{}(v).(safe_hasher.SafeHasher); ok {
