--- projects/sds/pkg/server/v2/v2_suite_test.go
@@ -1,13 +1,0 @@
-package sds_server_v2
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-)
-
-func TestSDSServer(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "SDS Server V2 Suite")
-}
