--- test/helpers/clients.go
@@ -1,15 +1,0 @@
-package helpers
-
-import (
-	. "github.com/onsi/gomega"
-	"github.com/solo-io/go-utils/kubeutils"
-	"k8s.io/client-go/kubernetes"
-)
-
-func MustKubeClient() kubernetes.Interface {
-	restConfig, err := kubeutils.GetConfig("", "")
-	ExpectWithOffset(1, err).NotTo(HaveOccurred())
-	kubeClient, err := kubernetes.NewForConfig(restConfig)
-	ExpectWithOffset(1, err).NotTo(HaveOccurred())
-	return kubeClient
-}
