+++ test/kube2e/gateway/gateway_suite_test.go
@@ -2,15 +2,29 @@ package gateway_test
 
 import (
 	"fmt"
+	"io/ioutil"
+	"net/http"
 	"os"
+	"os/exec"
 	"path/filepath"
+	"regexp"
+	"strconv"
 	"testing"
 	"time"
 
 	"github.com/solo-io/go-utils/log"
 
-	"github.com/solo-io/gloo/test/helpers"
 	"github.com/solo-io/gloo/test/kube2e"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
+	v1 "github.com/solo-io/gloo/projects/gloo/pkg/api/v1"
+
+	"github.com/gogo/protobuf/types"
+	clienthelpers "github.com/solo-io/gloo/projects/gloo/cli/pkg/helpers"
+	"github.com/solo-io/solo-kit/pkg/api/v1/clients"
+
+	"github.com/solo-io/gloo/test/helpers"
 
 	"github.com/solo-io/go-utils/testutils/helper"
 
@@ -39,26 +53,7 @@ func TestGateway(t *testing.T) {
 var testHelper *helper.SoloTestHelper
 
 var _ = BeforeSuite(StartTestHelper)
-var _ = AfterSuite(func() {
-	helpers.TearDownTestHelper(testHelper)
-})
-
-const (
-	helmValues = `
-global:
-  image:
-    pullPolicy: IfNotPresent
-  glooRbac:
-    namespaced: true
-    nameSuffix: e2e-test-rbac-suffix
-settings:
-  singleNamespace: true
-  create: true
-gloo:
-  deployment:
-    disableUsageStatistics: true
-`
-)
+var _ = AfterSuite(TearDownTestHelper)
 
 func StartTestHelper() {
 	cwd, err := os.Getwd()
@@ -76,7 +71,7 @@ func StartTestHelper() {
 
 	// Register additional fail handlers
 	skhelpers.RegisterPreFailHandler(helpers.KubeDumpOnFail(GinkgoWriter, "knative-serving", testHelper.InstallNamespace))
-	valueOverrideFile, cleanupFunc := kube2e.GetHelmValuesOverrideFile(helmValues)
+	valueOverrideFile, cleanupFunc := kube2e.GetHelmValuesOverrideFile()
 	defer cleanupFunc()
 
 	err = testHelper.InstallGloo(helper.GATEWAY, 5*time.Minute, helper.ExtraArgs("--values", valueOverrideFile))
@@ -87,9 +82,99 @@ func StartTestHelper() {
 
 	// TODO(marco): explicitly enable strict validation, this can be removed once we enable validation by default
 	// See https://github.com/solo-io/gloo/issues/1374
-	helpers.UpdateAlwaysAcceptSetting(testHelper.InstallNamespace, false)
+	UpdateAlwaysAcceptSetting(false)
 
 	// Ensure gloo reaches valid state and doesn't continually resync
 	// we can consider doing the same for leaking go-routines after resyncs
-	helpers.EventuallyReachesConsistentState(testHelper)
+	EventuallyReachesConsistentState()
+}
+
+func EventuallyReachesConsistentState() {
+	metricsPort := strconv.Itoa(9091)
+	portFwd := exec.Command("kubectl", "port-forward", "-n", testHelper.InstallNamespace,
+		"deployment/gloo", metricsPort)
+	portFwd.Stdout = os.Stderr
+	portFwd.Stderr = os.Stderr
+	err := portFwd.Start()
+	Expect(err).ToNot(HaveOccurred())
+
+	defer func() {
+		if portFwd.Process != nil {
+			portFwd.Process.Kill()
+		}
+	}()
+
+	// make sure we eventually reach an eventually consistent state
+	lastSnapOut := getSnapOut(metricsPort)
+
+	eventuallyConsistentPollingInterval := 7 * time.Second // >= 5s for metrics reporting, which happens every 5s
+	time.Sleep(eventuallyConsistentPollingInterval)
+
+	Eventually(func() bool {
+		currentSnapOut := getSnapOut(metricsPort)
+		consistent := lastSnapOut == currentSnapOut
+		lastSnapOut = currentSnapOut
+		return consistent
+	}, "30s", eventuallyConsistentPollingInterval).Should(Equal(true))
+
+	Consistently(func() string {
+		currentSnapOut := getSnapOut(metricsPort)
+		return currentSnapOut
+	}, "30s", eventuallyConsistentPollingInterval).Should(Equal(lastSnapOut))
+}
+
+// needs a port-forward of the metrics port before a call to this will work
+func getSnapOut(metricsPort string) string {
+	var bodyResp string
+	Eventually(func() string {
+		res, err := http.Post("http://localhost:"+metricsPort+"/metrics", "", nil)
+		if err != nil || res.StatusCode != 200 {
+			return ""
+		}
+		defer res.Body.Close()
+		body, err := ioutil.ReadAll(res.Body)
+		Expect(err).ToNot(HaveOccurred())
+		bodyResp = string(body)
+		return bodyResp
+	}, "3s", "0.5s").ShouldNot(BeEmpty())
+
+	Expect(bodyResp).To(ContainSubstring("api_gloo_solo_io_emitter_snap_out"))
+	findSnapOut := regexp.MustCompile("api_gloo_solo_io_emitter_snap_out ([\\d]+)")
+	matches := findSnapOut.FindAllStringSubmatch(bodyResp, -1)
+	Expect(matches).To(HaveLen(1))
+	snapOut := matches[0][1]
+	return snapOut
+}
+
+func TearDownTestHelper() {
+	Expect(testHelper).ToNot(BeNil())
+	err := testHelper.UninstallGloo()
+	Expect(err).NotTo(HaveOccurred())
+	_, err = kube2e.MustKubeClient().CoreV1().Namespaces().Get(testHelper.InstallNamespace, metav1.GetOptions{})
+	Expect(apierrors.IsNotFound(err)).To(BeTrue())
+}
+
+// enable/disable strict validation
+func UpdateAlwaysAcceptSetting(alwaysAccept bool) {
+	UpdateSettings(func(settings *v1.Settings) {
+		Expect(settings.Gateway).NotTo(BeNil())
+		Expect(settings.Gateway.Validation).NotTo(BeNil())
+		settings.Gateway.Validation.AlwaysAccept = &types.BoolValue{Value: alwaysAccept}
+	})
+}
+
+func UpdateSettings(f func(settings *v1.Settings)) {
+	settingsClient := clienthelpers.MustSettingsClient()
+	settings, err := settingsClient.Read(testHelper.InstallNamespace, "default", clients.ReadOpts{})
+	Expect(err).NotTo(HaveOccurred())
+
+	f(settings)
+
+	_, err = settingsClient.Write(settings, clients.WriteOpts{OverwriteExisting: true})
+	Expect(err).NotTo(HaveOccurred())
+
+	// when validation config changes, the validation server restarts -- give time for it to come up again.
+	// without the wait, the validation webhook may temporarily fallback to it's failurePolicy, which is not
+	// what we want to test.
+	time.Sleep(3 * time.Second)
 }
