+++ test/kube2e/util.go
@@ -11,7 +11,18 @@ import (
 	"github.com/solo-io/gloo/projects/gloo/cli/pkg/cmd/options"
 	"github.com/solo-io/go-utils/testutils/helper"
 	"github.com/solo-io/solo-kit/pkg/api/v1/resources/core"
+
+	"github.com/solo-io/go-utils/kubeutils"
+	"k8s.io/client-go/kubernetes"
 )
+
+func MustKubeClient() kubernetes.Interface {
+	restConfig, err := kubeutils.GetConfig("", "")
+	ExpectWithOffset(1, err).NotTo(HaveOccurred())
+	kubeClient, err := kubernetes.NewForConfig(restConfig)
+	ExpectWithOffset(1, err).NotTo(HaveOccurred())
+	return kubeClient
+}
 
 // Check that everything is OK by running `glooctl check`
 func GlooctlCheckEventuallyHealthy(testHelper *helper.SoloTestHelper, timeoutInterval string) {
@@ -35,19 +46,32 @@ func GlooctlCheckEventuallyHealthy(testHelper *helper.SoloTestHelper, timeoutInt
 	}, timeoutInterval, "5s").Should(BeNil())
 }
 
-func GetHelmValuesOverrideFile(values string) (filename string, cleanup func()) {
-	valuesFile, err := ioutil.TempFile("", "values-*.yaml")
+func GetHelmValuesOverrideFile() (filename string, cleanup func()) {
+	values, err := ioutil.TempFile("", "values-*.yaml")
 	Expect(err).NotTo(HaveOccurred())
 
 	// disabling usage statistics is not important to the functionality of the tests,
 	// but we don't want to report usage in CI since we only care about how our users are actually using Gloo.
 	// install to a single namespace so we can run multiple invocations of the regression tests against the
 	// same cluster in CI.
-	_, err = valuesFile.Write([]byte(values))
+	_, err = values.Write([]byte(`
+global:
+  image:
+    pullPolicy: IfNotPresent
+  glooRbac:
+    namespaced: true
+    nameSuffix: e2e-test-rbac-suffix
+settings:
+  singleNamespace: true
+  create: true
+gloo:
+  deployment:
+    disableUsageStatistics: true
+`))
 	Expect(err).NotTo(HaveOccurred())
 
-	err = valuesFile.Close()
+	err = values.Close()
 	Expect(err).NotTo(HaveOccurred())
 
-	return valuesFile.Name(), func() { _ = os.Remove(valuesFile.Name()) }
+	return values.Name(), func() { _ = os.Remove(values.Name()) }
 }
